# openapi-spec.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Travel Planer API",
        "description": "REST API for managing travel and trips and activities. Users can register, login with JWT authentication, create trips, add activities, and track budgets. ",
        "contact": {
            "name": "Nurbek",
            "email": "nurbek@example.com"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "tags": [
        {
            "name": "Activities",
            "description": "CRUD operations for trip activities"
        },
        {
            "name": "Trips",
            "description": "CRUD operations for travel trips, upcoming trips, and statistics"
        },
        {
            "name": "Authentication",
            "description": "User registration, login, and profile endpoints"
        }
    ],
    "paths": {
        "/api/trips/{tripId}/activities/{activityId}": {
            "get": {
                "tags": [
                    "Activities"
                ],
                "operationId": "getActivityById",
                "parameters": [
                    {
                        "name": "tripId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "activityId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ActivityResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Activities"
                ],
                "operationId": "updateActivity",
                "parameters": [
                    {
                        "name": "tripId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "activityId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateActivityRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ActivityResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Activities"
                ],
                "operationId": "deleteActivity",
                "parameters": [
                    {
                        "name": "tripId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "activityId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/trips/{id}": {
            "get": {
                "tags": [
                    "Trips"
                ],
                "operationId": "getTripById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/TripResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Trips"
                ],
                "operationId": "updateTrip",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateTripRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/TripResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Trips"
                ],
                "operationId": "deleteTrip",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/trips": {
            "get": {
                "tags": [
                    "Trips"
                ],
                "operationId": "getAllMyTrips",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "PLANNED",
                                "ONGOING",
                                "COMPLETED",
                                "CANCLEED"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TripSummaryResponse"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Trips"
                ],
                "operationId": "createTrip",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateTripRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/TripResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/trips/{tripId}/activities": {
            "get": {
                "tags": [
                    "Activities"
                ],
                "operationId": "getAllActivities",
                "parameters": [
                    {
                        "name": "tripId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ActivityResponse"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Activities"
                ],
                "operationId": "addActivity",
                "parameters": [
                    {
                        "name": "tripId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateActivityRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ActivityResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "operationId": "register",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "operationId": "login",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/LoginResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/trips/upcoming": {
            "get": {
                "tags": [
                    "Trips"
                ],
                "operationId": "getUpcomingTrips",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TripSummaryResponse"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/trips/stats": {
            "get": {
                "tags": [
                    "Trips"
                ],
                "operationId": "getTripStats",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/me": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "operationId": "getCurrentUser",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResponse"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "UpdateActivityRequest": {
                "required": [
                    "activityDate",
                    "cost",
                    "name",
                    "type"
                ],
                "type": "object",
                "properties": {
                    "name": {
                        "maxLength": 100,
                        "minLength": 0,
                        "type": "string"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "SIGHTSEEING",
                            "HOTEL",
                            "FLIGHT",
                            "RESTAURANT",
                            "TRANSPORT",
                            "OTHER"
                        ]
                    },
                    "activityDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "cost": {
                        "minimum": 0.0,
                        "exclusiveMinimum": false,
                        "type": "number"
                    },
                    "notes": {
                        "maxLength": 5000,
                        "minLength": 0,
                        "type": "string"
                    }
                }
            },
            "ActivityResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "SIGHTSEEING",
                            "HOTEL",
                            "FLIGHT",
                            "RESTAURANT",
                            "TRANSPORT",
                            "OTHER"
                        ]
                    },
                    "activityDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "cost": {
                        "type": "number"
                    },
                    "notes": {
                        "type": "string"
                    },
                    "tripId": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
            "UpdateTripRequest": {
                "required": [
                    "destination",
                    "endDate",
                    "startDate",
                    "status",
                    "title"
                ],
                "type": "object",
                "properties": {
                    "title": {
                        "maxLength": 100,
                        "minLength": 0,
                        "type": "string"
                    },
                    "destination": {
                        "maxLength": 100,
                        "minLength": 0,
                        "type": "string"
                    },
                    "description": {
                        "maxLength": 5000,
                        "minLength": 0,
                        "type": "string"
                    },
                    "startDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "endDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "budget": {
                        "minimum": 0.0,
                        "exclusiveMinimum": false,
                        "type": "number"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "PLANNED",
                            "ONGOING",
                            "COMPLETED",
                            "CANCLEED"
                        ]
                    }
                }
            },
            "TripResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "title": {
                        "type": "string"
                    },
                    "destination": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "startDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "endDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "budget": {
                        "type": "number"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "PLANNED",
                            "ONGOING",
                            "COMPLETED",
                            "CANCLEED"
                        ]
                    },
                    "userId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "activities": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ActivityResponse"
                        }
                    }
                }
            },
            "CreateTripRequest": {
                "required": [
                    "destination",
                    "endDate",
                    "startDate",
                    "title"
                ],
                "type": "object",
                "properties": {
                    "title": {
                        "maxLength": 100,
                        "minLength": 0,
                        "type": "string"
                    },
                    "destination": {
                        "maxLength": 100,
                        "minLength": 0,
                        "type": "string"
                    },
                    "description": {
                        "maxLength": 5000,
                        "minLength": 0,
                        "type": "string"
                    },
                    "startDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "endDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "budget": {
                        "minimum": 0.0,
                        "exclusiveMinimum": false,
                        "type": "number"
                    }
                }
            },
            "CreateActivityRequest": {
                "required": [
                    "activityDate",
                    "cost",
                    "name",
                    "type"
                ],
                "type": "object",
                "properties": {
                    "name": {
                        "maxLength": 100,
                        "minLength": 0,
                        "type": "string"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "SIGHTSEEING",
                            "HOTEL",
                            "FLIGHT",
                            "RESTAURANT",
                            "TRANSPORT",
                            "OTHER"
                        ]
                    },
                    "activityDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "cost": {
                        "minimum": 0.0,
                        "exclusiveMinimum": false,
                        "type": "number"
                    },
                    "notes": {
                        "maxLength": 5000,
                        "minLength": 0,
                        "type": "string"
                    }
                }
            },
            "RegisterRequest": {
                "required": [
                    "email",
                    "password",
                    "username"
                ],
                "type": "object",
                "properties": {
                    "username": {
                        "maxLength": 50,
                        "minLength": 3,
                        "type": "string"
                    },
                    "email": {
                        "maxLength": 100,
                        "minLength": 0,
                        "type": "string"
                    },
                    "password": {
                        "maxLength": 100,
                        "minLength": 6,
                        "type": "string"
                    }
                }
            },
            "UserResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "username": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "LoginRequest": {
                "required": [
                    "email",
                    "password"
                ],
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                }
            },
            "LoginResponse": {
                "type": "object",
                "properties": {
                    "token": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string"
                    },
                    "user": {
                        "$ref": "#/components/schemas/UserResponse"
                    }
                }
            },
            "TripSummaryResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "title": {
                        "type": "string"
                    },
                    "destination": {
                        "type": "string"
                    },
                    "startDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "endDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "budget": {
                        "type": "number"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "PLANNED",
                            "ONGOING",
                            "COMPLETED",
                            "CANCLEED"
                        ]
                    },
                    "activityCount": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "totalCost": {
                        "type": "number"
                    }
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter JWT token obtained from /api/auth/login endpoint. Format: Just paste the token (no 'Bearer' prefix needed).",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}